Flush FramedWrite before close
Ensure that the FramedWrite is fully flushed before attempting to
close the underlying I/O. Addresses comments by @najamelan in #11.

In addition, flush the I/O before we write to it in poll_write().
This is just in case the I/O does not make any progress / can flush
itself during a poll_write().