Fix length encoded frame parsing
The problem was that 'src' buffer was advanced too soon.
This caused a couple of bugs at least in 2 scenarios:

1. when encoded frame was smaller than 8 bytes.
2. when 'src' did not have enough bytes do decode a complete frame.

 ## Problem 1

Say we encode a frame [1, 2, 3].
The resulting buffer is [0, 0, 0, 0, 0, 0, 0, 3, 1, 2, 3] - with prepended
length header.
Then decode() takes this buffer, correctly parses len as 3, and then
does this comparison:

    if src.len() - U64_LENGTH

Given that we already advanced src buffer, src.len() now is 3 (11 - 8).
This results in uderflow: 3 - 8, which panics.

 ## Problem 2

Say we have framed a TCP stream where data is coming in chunks but not
necessarily in the size of our frames.
Say we call decode() for the first time with src = [0, 0, 0, 0, 0, 0, 0, 64, 1, 2, 3].
Obviously this is not a full frame: it's size is 64 but only first 3 bytes have arrived.
In this case decode() used to advance the src buffer, but still return None.
So when a next TCP packet arrives, decode() was called for the second time.
decode() then used to attempt to parse len header again. And this time it
would parse the bytes [1, 2, 3, ...] as a len header which is actually
our data.