Mark feature-gated items in docs
We’re making use of the [`doc(cfg(..))`][1] attribute to mark items in
the docs that are behind feature flags. The approach seems pretty
standard and is, for example, followed by the `futures-*` and
`async_std` crates.

To ensure this is tested we add an additional travis job for this. This
also checks the [`broken_intra_doc_links`][2] lint. To pass, we fixed
the links in the crate documentation in `lib.rs`.

[1]: https://doc.rust-lang.org/rustdoc/unstable-features.html#documenting-platform-feature-specific-information
[2]: https://doc.rust-lang.org/rustdoc/lints.html#broken_intra_doc_links